fix(discover): Handle custom events functions in Discover
This PR updates the select_entity function on the Discover dataset
to check for the presence of the isHandled and notHandled custom
functions. If they are present, we should attempt to use the Events
entity instead of the Discover entity.

If the Discover entity is still being picked (for example if other
transactions only columns are also present in the query), these functions
are mapped to Null.

This is equivalent to how we handle apdex and failure_rate in entity
selection and mapping.

This fixes SNUBA-22Q