feat: Remove special handling of promoted tags in errors replacer (#1440)
This removes all special handling of promoted tags in the errors replacer
and instead uses the values from tags.key and tags.value like our
unpromoted tags.

Unlike in the events storage, most of our promoted tags are low cardinality
columns now. This is causing a bug in ClickHouse -
ClickHouse/ClickHouse#16171.

Our workaround for now is to not use any of our promoted columns to query
for errors replacements; event replacements are unchanged.

This fixes SNUBA-22N in Sentry.