feat: Add merge table for Discover
Adds a migration to create the Discover merge table. This will eventually
replace the Events table, and all Discover queries where we want both errors
and transactions together will use this table.

Since the merge table requires the column names to match on all of the
underlying tables, we also need to add the following materialized
columns to transactions:
- type
- message
- title
- timestamp

Unfortunately, simply aliasing the columns will not work.