feat(schemas) Introduce column modifiers and stop making such modifier extend ColumnType (#1439)
While trying to refactor the column classes to support schemas with untyped columns (the SELECT clause of a subquery is a ColumnSet with untyped columns) I figured out I could not separate the hierarchy of DDL column types from the ones we use during query processing With ColumnTypeWithModifier in the way.

So I transformed Column modifiers into objects referenced by a column type instead of being column types themselves.
This way a modifier is an attribute of a column type (Nullable is an attribute of a String column instead of being a type itself.), which seems to me more in line with what modifiers represent.

Basically the interesting changes are in clickhouse/columns.py and migrations/columns.py. The rest is a gigantic codemod to adapt all schemas and migrations.