feat(joins) Allow nested query step 3: Make the query class a data source itself and allow subqueries (#1444)
With this step we are now able to create a query with subqueries in this form:

```
SELECT a, b, f(c)
FROM (
    SELECT a, b, g(x) as c
    FROM (
        SELECT a, b, c 
        FROM
           .....
   )
)
```
It does not do much so far since we cannot format it yet and we cannot use it for query processing yet. See test_nested.py

since the query class itself is a DatSource and it is able to produce a ColumnSet as a schema. This depended on the ColumnSet refactoring from last week so we can have an Any column type for schemas only (and not for migrations) since we cannot infer the type of all columns produced of a subquery.