feat: Add QueryPlanDelegator (aka Snüplicator)
This adds a basic skeleton of the QueryPlanDelegator, which is a type
of query plan builder that is able to delegate a request to one or
more other query plan builders which are executed against a single
query.

All query plan builders apart from the "primary" one (i.e. the
the one whose results are actually returned to the user) are run
asynchronously to ensure results can be returned before all of the
all of the subsequent queries are necessarily completed.

Currently the primary use case for this is to support the upcoming
transition of events queries to the errors table and Discover queries
to the new merge table. It will allow us to compare the correctness
and relative performance of various different query plans (and their
associated storages) before rolling any risky changes out to end users.

Inspired by Sentry's ServiceDelegator.