feat(query) Getting rid of aliases step 2b:  Replace SnubaResult column names with the names provided in the SelectedExpression objects after the query (#1458)
With this step column names in the results the user receives are those referenced in the SelectedExpression objects instead of being the aliases of the related expressions in the AST.
This fully decouples the aliases we maintain in the AST from the column names in the results. Thus we can first replace aliases with internal names and potentially later, remove the concept of aliases altogether. Already using internal names (like _snuba_whatever for the aliases in the AST will remove the alias shadowing problem entirely.

Right now nothing really changes because the name of a SelectedExpression object and the alias in the expression it contains are the same string.