feat: Add basic query pipeline and query pipeline builder class
This PR adds the minimal QueryPipeline and QueryPipelineBuilder classes,
and two subclases - SingleQueryPlanPipeline and MultipleQueryPlanPipeline.
Right now, this doesn't really do anything new - the single query plan
pipeline just executes the query plan provided and returns the result.
The multiple query plan pipeline is not yet used (or complete) and will
allow multiple query plans to be executed against a single query in order
to experiment with and test different query plans in parallel.