feat: Implement QueryPlanDelegator
This is mostly a port of https://github.com/getsentry/snuba/pull/1454
but QueryPlanDelegator now builds a query pipeline which may build plans
from more than one query plan builder.

MultipleQueryPlanPipeline delegates a request to one or more query plan
builders which are executed against a single query. All query plans
apart from the "primary" one are built and run asynchronously and
results can be returned before all of the of the subsequent ClickHouse
queries are necessarily completed.

The primary use case for this is to support the upcoming transition of
events queries to the errors table and Discover queries to the merge
table. It will allow us to compare the correctness and relative
performance of various different query plans (and their associated
storages) before rolling any risky changes out to end users.