feat: Add errors_ro storage
The errors readonly storage is the errors equivalent of the events
readonly storage, and will be used in the same scenarios where events_ro
is now - i.e. when `enable_events_readonly_table` is set and the request
request is not required to be "consistent", allowing us to use other
replicas than the first one.

The errors_dist_ro table is already being created through the migration
system.