feat(migrations): DropIndex uses IF EXISTS operator
We should use `DROP INDEX IF EXISTS` instead of bare `DROP INDEX` so
that we can reverse migrations even if an index wasn't successfully
added. This came up recently in the migration 0008_transactions_add_timestamp_index
- if the original AddIndex command failed the migration will be marked "IN
PROGRESS" but not have the index present. The user would be stuck and
unable to run either the forwards or backwards migrations.