feat(join): Composite query formatter step 2: Remove dependency of Query on RelationalSource (#1467)
RelationalSource was our from clause representation from an ancient time when storages, entities, composite queries etc.
RelaitonalSource is a structural part of the data model (it is used by TableSchemas to define the schema of the table to create).
This did not fit really well with the new query processing, where the query is generated starting from the structural elements of the data model, but where the query does not reference them.
This replaces RelationalSource (i the from clause) with a dedicated structure that is not used to build the schema. This structure is Table and extends SimpleDataSource. This allows us to provide attributes to this dat asource that are proper of the from clause but that are not part of the schema (like sample rate and final).

This steps makes it much easier to move ahead with the formatter.