feat(join): Composite query formatter step 3: Move final ad sample attributes from query to from clause. (#1468)
final and sampling rate (according to clickhouse) are properties of the from clause, so moving them there (in the Table data source) now that we do not use RelationalSource anymore for the from clause.
This simplifies a lot formatting since we can format the from clause using only the data source. It also allows us not to have a sampling rate on composite queries (which do not make sense).