feat(join): Composite query formatter step 4: Allow FormattedQuery to be nested (#1471)
In order to format nested queries (which include subqueries and joins) the FormattedQuery abstraction needs to be adapted so that we can actually format nested query to string and we can represent them as a dictionary for tracing (we use the dictionary representation for tracing in discover).

expand FormattedQuery to allow fully nested structures (which now can be exported as nested sequences or strings)
moves around files. create a snuba.clickhouse.formatter module with all the formatters submodules
refactors the query formatter to separate query formatting from data source formatting.