feat(join): Composite query formatter step 5: Add a formatter for the JoinClause structure. (#1472)
Adds a formatter for the JoinClause tree. This generates a FormattedNode (introduced in #1471) from the join tree which is expressed as JoinNode.
Specifically:

adds a visitor to navigate the JoinnClause tree (here it is possible since all nodes are in the same module, so on circular ddependencies)
implements it in the query formatter.
adds tests.