feat(join): Composite query formatter step 6: Put everything together and format joins and subqueries combined. (#1473)
This is the last step and it allows to format CompositeQuery. It allows composite queries to be formatted, which means:

- subqueries
- queries containing simple joins
- joins of subqueries
- queries over joined queries
- all the combinations of the above.

To achieve this, it breaks down the format_simple_query method into a method per clause so they can be reused when formatting simple queries and composite queries. It also uses the JoinVisitor introduced in #1472 to format the FROM clause when in the form of a join.

It adds tests for the complex cases in test_query_format.py