feat(snql) Add a number more features to the language
This adds support for the SnQL language except JOIN clauses. It also adds some
more tests and fixes a number of bugs. One of the biggest changes (codewise) was
to enforce spaces between major clauses (`MATCH (e:events)SELECT` is no longer
valid). Also rename COLLECT to SELECT and move it in front of the WHERE clause
to better match SQL format.