fix(release): Add missing GITHUB_API_TOKEN variable to craft/publish
This fixes the recurring failure when merging the release branch to master when releasing Snuba. The failure happens because we have branch protection enabled and the default `github.token` is not an admin on the repo to by-pass these checks. This token was already added on Sentry and Relay repositories so this was not an issue there.