feat: Implement PipelineDelegator and ThreadedFunctionDelegator(#1465)
Introduces the ThreadedFunctionDelegator, which executes one or more functions
against the input provided. The list of callables that are executed are determined
by a selector function. The "primary" one is returned and the rest are executed in
threads. The results of all functions are made available via the callback function if
one is provided.

The MultipleConcurrentPipeline uses the ThreadedFunctionDelegator in order to
build and executes requests against multiple pipelines in parallel.

The primary use case for this is to support the upcoming transition of events queries
to the errors table and Discover queries to the merge table. It will allow us to compare
the correctness and relative performance of various different pipelines (and their
associated query plans and storages) before rolling any risky changes out to end users.