feat(joins) Support composite queries in db_query step 2: Support Composite queries in db_query (#1504)
The biggest obstacle for making db_query take in a CompositeQuery was that it needs to count the Clickhouse columns referenced by the query. All of them.

This uses the visitor introduced in #1503 to scann all the columns in the data source of the query and tell apart references to query exposed by subqueries from real columns that are actually referenced.