feat(errors): Make `level` nullable
The `level` column should be a nullable string because it is a promoted
tag. Non nullable promoted tags are problematic because the insertion will
fail if the tag is absent in the payload. If we attempted to use empty string
as a default instead, it would cause inconsistent query results.