test: Simplify discover entity selection tests
Since we are only really concerned with testing entity selection here,
we can just assert that the entity in the query is what we expect and
skip running the rest of the pipeline. The actual table name is subject
to change since we will be gradually migrating from the `sentry_local`
to `errors_local` and `discover_local` tables.