feat(discover): Add the ability to run queries on merge table
This uses the pipeline delegator to simultaneously run queries to the
Discover entity on the Discover merge table, in addition to the main
events table. This does not yet change any of the results being returned
to users.

The percentage of queries that will also run on the alternate storage
can be specified via the "discover_query_percentage" configuration. The
default is 0.

If more than one storage has been run, we compare the result data from
both storages and increment a metric recording whether they are the same
or different.