fix(split): Use standard column aliases in splitter
This is a quick and kinda dirty fix for the event_id alias shadowing
issue.

We changed the query parser to mangle column aliases so that we don't
shadow the actual column names and can reference them in processors,
however we forgot to also use the mangled aliases in the column
splitter, rather than apply the original column names.

This should fix SNUBA-20.