fix(split): Do not perform column splitting on aggregations
Do not do column splitting if we are attempting to select anything that
is not a column or a literal. This is to avoid column splitting when
there are aggregations present in the query, which can cause errors.

Unfortunately this means queries that reference selected tags no longer
use the column splitter as these have been translated to function calls.

Fixes SNUBA-20G