Revert "perf(transactions): Turn on column splitter (#1513)" (#1520)
This reverts commit c035fe6f2aec554459f8a4d95e138a209e48289d.

This caused problems when there are aggregations (but no group by clause) in the query.
In that case the splitter should be skipped but is not. For some reason, this only commonly
happens in production on the transactions dataset.
Reverting first since fixing these turned out to be more involved than I originally thought.