fix(transactions) Fix query error in uuid processor (#1521)
If the client passes in an invalid event id, then the parser was leaving the
event id unchanged but still stripping the replaceAll(... from the left side.
This would cause errors in Clickhouse. Instead, if the event id is not a valid
UUID, bail out of the processor entirely and leave the expression unchanged.