feat(joins) Composite query execution pipeline step 3: Introduce the CompositeQueryPlanner (#1510)
This is the bulk of the processing of a composite query. It introduces the CompositeQueryPlanner which takes a logical CompositeQuery and produces a CompositeQueryPlan.

Like the simple ClickhouseQueryPlan, the CompositeQueryPlan contains the Clickhouse composite query generated by translating the logical composite query. It also contains all the query processors that have to be executed by the query execution strategy.

The main differences from the ClickhouseQueryPlan are in the translation and in the query processors.
The translation of the composite query is simpler (there are no translation rules) though the complexity comes from the from clause. This is nested, it can contains simple queries, joins of subqueries, or nested composite queries. So we need to visit all the nodes of this tree and perform the translation of simple queries according to the translator provided by their entity.

There are no composite query processors at this stage (we are likely to introduce some later when dealing with optimizations). Though all query processors of all subqueries have to be executed.
These are executed by the query execution strategy (as we are likely to end up splitting the composite query as well), so the composite query plan needs to keep track of all of them and to associate each subquery with its own processors.

Also one verbose test is added that plans a composite query over a simple query, and a composite query over a join of two simple queries.