feat(snql) Discover tests using SnQL (#1517)
Have the discover tests use SnQL. This exposed a few missing cases, particularly
ARRAY JOIN. Add support for the ARRAY JOIN clause. The current language also
supports an implicit transformation on array columns, converting conditions to
be arrayExists or arrayAll functions if the condition is on an array column.

This is ambigious and hidden behaviour that we are trying to avoid in SnQL.
Instead, add a special function the client can use as syntactic sugar for the
same condition, forcing the client to explicitly declare when they are using a
condition on an array column. There are two discover tests that rely on this
behaviour, so xfail them in snql for now.