Add join/subquery support to MATCH clause (#1524)
Add support for join clauses in the MATCH clause, as well as support for sub
queries in the match clause. This required writing a complicated linked list
algorithm for the join clause, to ensure the joins were possible and to order
them correctly for clickhouse. It also required changing the post process query
functions to support composite and logical queries.