feat(joins) Generate the sets of equivalent columns across entities (#1535)
This PR provides a function to extract the sets of equivalent columns between all the entities inn a JoinClause. The PR on top of this will use those sets to generate the missing conditions.
There are two sources to provide the equivalences. One is obviously the join key. The other is a static set of equivalences that are declared in the entity (added in this PR). This is not part of the join key and covers cases like the transaction_name columns.