feat(join) Subquery generation step 1: Basic subquery creation (#1539)
This is the first step in turning a query that joins multiple entities into an optimized join of subqueries.
This step is required because each subquery is processed independently following the usual query processors, translation and storage selection so they have to contain all the necessary structure that has to be moved from the external query.

This step simply initializes the subqueries and naively adds all the columns from the external query into the select clause of the subqueries. The result should be a runnable query (subscriptable references apart), though no query processor could do anything useful. The next step would start pushing down more meaningful expressions.