build: Add VERSION and related bump scripts (#1545)
Today we needed to release a patch release right after a CalVer release. Since there is no hard-coded version in Snuba anywhere, Craft got confused when it ran the pre-release script as it did not change anything (it only updates the license date which limits this to 1 release per day). I think it is good practice to hard-code the version in at least `setup.py` even if it is not referenced by anything yet. This also resolves our problem with releases.