fix(discover) Literal caching is not typed
This actually fixes two bugs. The first is that we were incorrectly caching
Literals with equal int and float values as the same thing. This would break
the subsequent Clickhouse query since the typing wouldn't match.

The second is that the SnQL parser was not correctly handling the HAVING clause.