feat(join) Push down expressions to subqueries step 2: Skip aggregation function. (#1553)
PR #1552 introduces a visitor that prepares query expression to be pushed down into subqueries.
That did not address aggregation functions. These are tied to the relative GROUP BY clause. We cannot push down an avg(duration) and do the group by in the main query.
Each aggregation needs to be at the same level where we do the group by.

Since we are not yet pushing down group by (which will be needed only much later if we group by the join key) this ensures we keep the aggregation functions in the main query.