feat(joins) Push down expressions step3: Fit an alias generator in the branch cutter (#1556)
We cannot rely on all expressions to have an alias to push them into subqueries.
In order to push down an expression we need to be able, in the general case, to reference it from the main query. Thus it needs to have an alias to ensure it can always be referenced.
This means we need to generate aliases for the expressions we push down and that do not have an alias (like conditions).

This adds the concept of alias generator (which is a generator) which is provided to the branch cutter and that produces a new unused alias every time we iterate on it.
The generator is provided externally, so we can have one per query making it easy to avoid duplicates.