feat(backfill): Update backfill to use insert into select
The backfill script now runs an INSERT INTO SELECT statement on
each partition to copy data from the events to errors table.
Once the migration is done OPTIMIZE TABLE will be run on each
part if the ClickHouse installation is single node. For distributed
tables, this should be performed manually on each node.