fix(snql) Mangle aliases in subqueries
There is an issue with mangling aliases in subqueries. An outer query is
expecting its columns to match the select statement from the inner query.
I assumed this meant the outer query should match the SelectedExpression,
but that isn't correct in a subquery since the SelectedExpression is
essentially ignored. Instead, the outer query columns need to match the
aliases of the inner query. Since those aliases get mangled, we have to also
mangle the column names of the outer query to match that mangling.