fix(joins) Ensure subqueries generated during the optimization phase have proper aliases (#1565)
When we have a join inside a subquery, we need to ensure the branches that are cut and replaced by references can be referenced by the external query as well.
Example:

SELECT avg(f_id) FROM (SELECT f(a.id) as f_id FROM ... join ...) a

After cutting f(id) and pushing it down we replace it with a column that references f(id) in the subquery. Though we need to ensure that f(id) can still be referenced by the external query in avg(f_id).

So when we push down f(id) we will preserve the alias in the column we create. That becomes:

SELECT avg(f_id) FROM
   SELECT a.f_id as f_id
        FROM
           (SELECT f(id) as f_id FROM ...) a
           JOIN
           (SELECT .....) b