Allow a plan builder to rank and return multiple plans (#1566)
When we execute join queries, each entity provides a query plan builder, this picks a storage and build a QueryPlan for that subquery.
If an entity has multiple viable storages for a query on different StorageSet, the join planner needs to know about all of them (ranked from best to worst) to ensure we pick storages on the same storage set for all entities.
If that does not happen the query would fail because the storages may be on different clusters.

This changes the interface of the QueryPlanBuilder to build multiple plans and either return the best or return the whole list. Single query pipelines use the best directly, the join planner has access to all viable options.