fix(snql) Remove extensions from SnQL queries
When a SnQL query is made, the extension fields should already be part of the
query so we don't need the extensions parsed explicitly. Also remove the
timeseries extension call from the query file, since from_date and to_date
were only being used to populate num_days, which is not used anywhere.

I tested this locally with the querylog consumer and the sentry tests and saw
no errors.