feat(errors): Connect the events entity to errors storage (#1581)
The events entity now runs some queries simultaneously on the errors storage
using the pipieline delegator. The purpose of this is to start testing and
comparing results to events on the two storages.

The percentage of errors that are run on both storages is determined by the
`errors_query_percentage` config. Default is 0.