Explicitly set quantity to 1 (#1579)
PR #1455 added the quantity column to the sessions raw table with a DEFAULT = 1.
Turns out that on distributed tables we cannot trust the default value. When querying the distributed table the quantity is still 0.
This makes it impossible to apply the migration in #1487 as it relies on the quantity to be 1 for individual sessions.

In order to fix that we need to manually set the quantity value to 1 instead of leaving it empty.