feat(join) Semi join optimization (#1592)
A query like this:

SELECT max(t.duration) FROM transactions t INNER JOIN span s
Does not need to build the entire hashtable as we do not reference anything on the right side. Thus in this case we can switch it to a SEMI join, which means clickhouse will return only the first row it can find from the right side instead of all of them. Which means fewer chances of out of memory.

This PR switches the join clause to semi join if the query does not need anything from the right side.