feat(joins) Move the query body from the query class to the request class  (#1601)
Composite queries do not have a body field (rightfully). The raw query body should be associated to the request instead of the query.
This prevents any stage from the query processing to depend on the raw body (as they only have access to the Query class) and still preserves the raw body accessible to the code that orchestrate the query execution pipeline. This logic needs it mostly for tracing and logging (see the querylog).