feat(snql) Add condition validation
In the legacy API, entities had extensions that would automatically add the
conditions that we required on Clickhouse. SnQL will not provide conditions
automatically, but we still need to verify that a query has the required
conditions.

Add required conditions to the entities, and a check in the SnQL parser to find
a top level condition for those columns. At the moment we expect the bare column
in the condition, which seems reasonable considering the current extensions.