feat(migrations): Add --dry-run option to migrations tool (#1615)
Add the --dry-run option to the migrations run and reverse commands.
This outputs the SQL that will be run on the ClickHouse nodes. Non
SQL operations (like Python scripts) cannot be previewed with --dry-run.

The --dry-run option also works with distributed and replicated tables,
which are not yet otherwise supported via the automated migration tool.
This is determined by the CLUSTERS settings configuration - if a particular
cluster has single_node set to false then the DDL that will be output will
be the distributed and replicated variant.