test(multiprocess) Add some metrics to identify when the consumer is stuck waiting  (#1620)
When running the multi process consumer with the INFO log level, we do not know when and if the consumer gets paused. This made the investigation of an incident harder.
This adds logging in two cases:

the consumer was paused for more than 10 seconds. Not logging every pause events because under some circumstances that can happen more often than a batch being written.
the consumer waited for results from the worker processes longer than 20 seconds without making progress.
It also records the size of the input batch to identify spikes in the size of the input batch that could justify a longer pause.