feat(migrations): Add a command to add a new node to a cluster
The add-node command supports running local or distributed migrations
on a brand new ClickHouse node. This is designed to help with adding
shards and replicas to existing clusters. The migrations need to be run
(with add-node) first before the machine can be added to the cluster,
otherwise queries will fail due to missing tables.

The command runs all SQL operations relating to the storage set/s
provided. Non SQL operations will be skipped.