fix(mypy) Fix type hints in snuba.consumers
I had to change the type of the KafkaBrokerConfig to match how the processor is
using that value. I spent a bunch of time trying to avoid the type:ignore on the
process_message, but TypedDicts just don't work nicely with typing. One solution
that might have worked is having a huge function that took a Mapping and
manually created each Event with every field explicitly defined. Or refactor
the MessageProcessor class and how we use it.