feat(multirprocess consumer) Log sigchld from the parallel transform step. (#1644)
The likely reason for the multi-process multi-storage consumer getting stuck in production is that one or more children process are terminated for any reason. A possible reason can be the OOM killer.
When a child process is terminated the multiprocessing pool hangs.

The idea is to cause a termination of the consumer should that happen, and intercepting sigchld appear to be the only viable way. Ideally we would not impose timeout or rely on private fields of the multiprocessing pool.

This step will log such an event without doing anything to ensure this signal is not sent under normal circumstances.