feat: Add query path for using errors and discover storages
This enables incremental rollout by referrer of the errors and discover
storages.

settings.ERRORS_ROLLOUT_BY_REFERRER and settings.ERRORS_ROLLOUT_ALL
are used to control which referrers (or all) are using the new storages.

There is also  a kill switch to allow a quick revert if required. In Snuba config,
"errors_rollout_killswitch" can be set to 1 to kill the entire rollout and
revert back to using the original events table for all Events and Discover queries.