feat: Default events writable storage is toggled by settings
Whilst we are in the process of rolling out errors storage, the
default writable storage can be toggled via settings.

The same killswitch used by the query path "errors_rollout_killswitch"
can be used to quickly revert back from the errors to events table.

This mostly affects the test endpoints which are used by Sentry and Sentry CI
however would also affect anyone who happens to be using the SnubaEventStream
rather than KafkaEventStream implementation of the eventstream in Sentry.