feat(errors): Default replacement and commit log topic names match events (#1675)
Update the default replacement and commit log topic names to match those
of the events storage. This allows us to safely swap the events consumer
and replacer for the errors consumer and replacer, for environments
like dev and onpremise where the multistorage consumer isn't being used
as part of the rollout strategy and these defaults are relied on.