feat: Attempt to assign discrepancies in Snuplicator results to a category
This is an attempt to categorize the discrepancies observed between the events
and errors storage, and quantify the impact of each category.

This implementation is primarily based on the premise that each "referrer"
sent from Sentry tends to be associated with a specific type or shape of query,
and knowing this context about the kind of query we are dealing with and where
it comes from can give us some better insights into what might have gone wrong.

This initial implementation of the "assign_reason_category()" function looks for
one main class of mismatch - mismatched aggregate values when everything else
remains the same. I picked this first because I suspect it accounts for the
lion's share of the discrepancies that we are currently observing. Later, we
should try to add more expected scenarios here though.

The "reason" is tagged in both our Sentry warnings and Snuplicator metrics so
we can break down our data by this additional dimension.