feat(snuplicator): Allow customizing % of queries run by referrer
The purpose of this change is to help understand the impact of caching
(both caching in Snuba as well as ClickHouse's internal caching) on query
timings, and makes the reported timings from the sentry_local and errors_local
tables more comparable.

Only running a low sample (1%) of all types of queries on the errors storage
can result in a lower cache utilization rate compared with events. This change
allows us to roll out a larger percentage of certain types of queries without
massively increasing the overall load on the system if we just turned the sample
rate up to 100%.