feat(migrations): Support global Python migrations
The purpose of this change is to support more complex migrations that cannot
solely be defined as a series of SQL statements. It is required in order to
support the upcoming migration that will backfill the new errors table from the
events table.

This change splits the generic MultiStepMigration class into two separate
migration classes which serve distinct purposes.

1. The ClickhouseNodeMigration should be used for "standard" ClickHouse migrations.
When writing a migration of this type, the list of local and dist operations must be provided,
and the migration runner orchestrates executing these on the correct set of ClickHouse nodes.
This is pretty much the same as how today's MultiStepMigration behaves, however it's now
restricted to SQL operations only. This is the one to be used in most standard cases of
schema migrations.

2. The GlobalMigration should be used for most data migrations. Only Python operations
can be provided in a global migration. The operations are simply executed once (not once
per node). The migration operation itself must establish the relevant ClickHouse connections
where necessary.