feat(snuplicator): Record cache_hit and is_duplicate with Snuplicator metrics
Records whether cache_hit and is_duplicate were triggered for any queries run
on both events and errors/discover storages.

Unfortunately this solution involves passing the query metadata all the way through
the pipeline so we have access to it in the callback function. I was tempted to make
the query metadata an attribute of a Request to simplify this, but that's probably
not the right spot for it since it doesn't have much to do with the request as
received from the client.