feat(migrations): Support global Python migrations (#1706)
The purpose of this change is to support more complex migrations that cannot
solely be defined as a series of SQL statements. It is required in order to
support the upcoming migration that will backfill the new errors table from the
events table.

This change splits the generic MultiStepMigration class into two separate
migration classes which serve distinct purposes.

1. The ClickhouseNodeMigration should be used for "standard" ClickHouse schema migrations.
When writing a migration of this type, the list of local and dist operations must be provided,
and the migration runner orchestrates executing these on the correct set of ClickHouse nodes.
This is pretty much the same as how today's MultiStepMigration behaves, however it's now
restricted to SQL operations only. This is the one to be used in most standard cases of
schema migrations.

2. The CodeMigration should be used for most data migrations, or anything that is more
complex than plain SQL. Only Python operations can be provided in a code migration.
The operations are simply executed once globally (not once per ClickHouse node). If any
ClickHouse connections are required in the course of the migration, the migration operation
is responsible for establishing those itself.