feat(prewhere): Allow modified prewhere candidates if query runs with final
Adds a mechanism to prevent specified columns from being included in the prewhere
clause if a query runs with final.

This helps us avoid a fairly obscure ClickHouse bug that causes an error when
a query contains a prewhere on a low cardinality, nullable column and that
query is run with final.

Since the new errors storage uses this column type more often than the events
table that it replaces, this may start to come up more often. It would also allow
us to safely add the `release` column to the list of prewhere candidates as long
as final is not used.

Fixes SNUBA-26G