feat(migrations): Enforce presence of Snuba columns in database (#1723)
Currently our tests ensure that all database columns must be present in
the corresponding Snuba schema, but not vice versa. This is problematic
since in many cases columns must be added to the database first before they
are referenced in code. A missing database column can cause replacers
or other processes that simply operate on the full list of columns as defined
in the schema to crash.

This flips the check around. The database may have extra columns than the Snuba
schema, but all columns in Snuba's schema must be present in the database.

This came up in https://github.com/getsentry/snuba/pull/1711#issuecomment-784416373

Depends on https://github.com/getsentry/snuba/pull/1718