feat(query) Add commutative condition matcher (#1729)
We have to match conditions often that could have literal on either side of the expression. This is specifically true with SnQL that does not impose the literal to be on the right hand side.

This adds a helper function to generate a proper pattern to match conditions in a commutative way. It will be useful for promoting the trace_id and span_id contexts.