perf(contexts) Promote span_id and trace_id contexts on the transactions storage. (#1730)
trace_id and span_id cannot be promoted like any other tag or context since the promoted columns are not strings.
Moreover even casting the result into string is not enough because the trace_id column is a proper UUID while the string in the contexts column does not have hyphens. The span_id is stored as an integer instead while the context contains a hex encoded string.

This requires a dedicated processor that promotes these columns when they are inside conditions and ignore them in the SELECT clause or group by so that the client still sees the expected format.

This is the first step. It just takes care of conditions.
We can follow up by promoting the column everywhere and converting them back to what the user expects in select/groupby. This way we would not be loading the contexts column at all when possible.