feat(migrations): Add a command to add a new node to a cluster (#1627)
The add-node command supports running local or distributed migrations
on a brand new ClickHouse node. This is designed to help with adding
shards or replicas or distributed hosts to existing clusters. The migrations
need to be run (with add-node) first before the machine can be added to the
cluster, otherwise queries will fail due to missing tables.

The command runs all SQL operations relating to the storage set/s
provided. Non SQL operations will only be run if an explicit `new_node_func` is
provided.

The test added here is the first test (but won't be the last) that tests replicated
table functionality. This means it requires Zookeeper to be configured as well as
ClickHouse macros to be set. I've added this configuration to the docker test
command but we may want to consider also adding this option to Sentry's devservices.